tests: refactor Matcher related integration tests
