+++ tests/lib.rs
@@ -1,3 +1,4 @@
 extern crate actiondb;
 
 mod file;
+mod matcher;
