+++ tests/matcher/ssh_ok.pattern
