+++ tests/matcher/ssh_wrong.pattern
